Fix server crash
This fixes an issue where a player who dropped all of their weapons
except the ghost will crash the server when RemoveGhost() gets called on
them. The bug happens because the m_hLastWeapon property of such a
player will point to a weapon that they don't hold, so the subsequent
attempt to set their m_hActiveWeapon to it will trigger a crash. This
can happen if a knifeless player (ie. a support) has dropped all of their guns, and used
all of their nades, and then caps the ghost.

Bug report and reproducible case provided by bauxite (thanks!)