Fix server crash on late load
The ghost entref variable is default-initialized to 0, and later in
RemoveGhost, its validity is checked with IsValidEdict. But 0 passes
IsValidEdict, because it is the server's edict. If the plugin loads late
and misses the OnGhostSpawn event which sets the ghost reference to
something nonzero, a RemoveGhost call will send the "Kill" input to entity 0 (the server itself),
which causes a crash.